+++ exporter/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ExporttoolConfig(AppConfig):
+class ExporterConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'exporttool'
+    name = 'exporter'
