+++ exporter/templates/exporttool/filterhelp.html
@@ -1,4 +1,4 @@
-{% extends 'exporttool/main.html' %}
+{% extends 'exporter/main.html' %}
 
 	{% block content %}
