+++ exporter/templates/exporttool/index.html
@@ -1,5 +1,5 @@
 
-{% extends 'exporttool/main.html' %}
+{% extends 'exporter/main.html' %}
 
     {% block content %}
