+++ exporter/templates/exporttool/main.html
@@ -9,7 +9,7 @@
 <body>
 
 	<!-- BLOCK TAGS! -->
-	{% include 'exporttool/navbar.html' %}
+	{% include 'exporter/navbar.html' %}
 
 	{% block content %}
