+++ fluxx_exporter/urls.py
@@ -19,5 +19,5 @@ from django.urls import include, path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('exporttool.urls'))
+    path('', include('exporter.urls'))
 ]
